requirements.txt
click==8.1.7
loguru==0.7.2
pytest==8.3.3
python-dotenv==1.0.1
celery==5.4.0
google-generativeai==0.8.2

llama_parse==0.5.6
llama-index==0.11.23
llama-index-readers-file==0.3.0
llama-index-readers-json==0.2.0
llama-index-vector-stores-qdrant==0.3.0
llama-index-postprocessor-cohere-rerank==0.2.1
llama-index-postprocessor-rankgpt-rerank==0.2.0
llama-index-postprocessor-rankllm-rerank==0.2.0

schedule==1.2.2
theflow==0.8.6
langchain==0.3.9
docx==0.2.4
html2text==2024.2.26
PyMuPDF==1.24.10
unstructured==0.15.13
openpyxl==3.1.5
python-docx==1.1.2
Redis==5.1.0
torch==2.4.1
transformers==4.45.1
python-pptx==1.0.2
Pillow==10.4.0
qdrant_client==1.11.3
fastapi==0.115.0
uvicorn==0.31.0
sqlmodel==0.0.22
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
minio==7.2.9
mmengine==0.10.5
elasticsearch==8.15.1
pyjwt==2.9.0
passlib[bcryt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
flower==2.0.1
uvicorn[standard]==0.31.0
fastapi-utils==0.7.0
langfuse==2.51.4
tiktoken==0.7.0
json_repair==0.30.0
wordcloud==1.9.3
reportlab==4.2.5
embedchain==0.1.125
crewai==0.80.0
crewai-tools==0.14.0
qdrant-client[fastembed]
boto3
s3fs
pymysql